ml/expense_analyzer/requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
receipt-ocr>=0.3.1
python-dotenv>=1.0.0
